refactor(pawnMemoryComp): 清理废弃代码并同步更新 API 调用
- 移除 PawnMemoryComp 中所有用于向下兼容的废弃 API 封装（如 ShortTermMemories, AddMemory 等），使其仅保留作为旧存档兼容类的作用。
- 将原本散落在外的 CompProperties_PawnMemory 合并至 PawnMemoryCompAlias.cs 文件中，并将文件改名为PawnMemoryComp.cs。
- 全局同步更新调用链：AI 集成、管理器及 UI 层现已直接调用 FourLayerMemoryComp 的原生属性与方法（如 AddActiveMemory, SituationalMemories.Count, DecayActivity）。
- 将必要的记忆上下文查询方法（GetMemoryContext, GetRelevantMemories）上移至 FourLayerMemoryComp 基类，等待后续进一步解耦。
- 注释及移除 UI 和补丁中不再维护的旧功能代码（ArchiveAll, GetMemorySummary）。
